refactor: :truck: Moved traverse functions to new file in deepembedding
