+++ parsley/shared/src/main/scala/parsley/internal/deepembedding/frontend/PrecedenceEmbedding.scala
@@ -3,22 +3,8 @@ package parsley.internal.deepembedding.frontend
 import parsley.internal.deepembedding.backend, backend.StrictParsley
 import parsley.internal.deepembedding.ContOps
 import parsley.internal.deepembedding.ContOps.{result, suspend, ContAdapter}
+import parsley.internal.deepembedding.Traverse.{traverse, traverse_}
 import parsley.internal.deepembedding.frontend.LazyPrec.{Atoms, Level, traverse_ => traversePrec_}
-
-object TraverseHelper {
-  def traverse[M[_, +_]: ContOps, R, A, B](xs: List[A])(f: A => M[R, B]): M[R, List[B]] = xs match {
-    case Nil => result(Nil)
-    case x :: xs => for {
-        y <- f(x)
-        ys <- traverse(xs)(f)
-    } yield y :: ys
-  }
-
-  def traverse_[M[_, +_]: ContOps, R, A, B](xs: List[A])(f: A => M[R, B]): M[R, Unit] = xs match {
-    case Nil => result(())
-    case x :: xs => f(x) >> traverse_(xs)(f)
-  }
-}
 
 private [parsley] final class Precedence[A](table: LazyPrec[A]) extends LazyParsley[A] {
   override protected def findLetsAux[M[_, +_]: ContOps, R](seen: Set[LazyParsley[_]])(implicit state: LetFinderState): M[R,Unit] = {
@@ -35,12 +21,12 @@ private [parsley] final class Precedence[A](table: LazyPrec[A]) extends LazyPars
   } yield new backend.Precedence(strictPrec)
 
   private def buildStrictPrec[M[_, +_]: ContOps, R, X](lp: LazyPrec[X])(implicit lets: LetMap): M[R, backend.StrictPrec[X]] = lp match {
-    case Atoms(atoms) => TraverseHelper.traverse(atoms)(_.optimised[M, R, X]).map(backend.StrictPrec.Atoms(_))
+    case Atoms(atoms) => traverse(atoms)(_.optimised[M, R, X]).map(backend.StrictPrec.Atoms(_))
     case lvl @ Level(lower, lOps) => {
       type Op = lOps.f.Op[lvl.In, X]
       for {
         strictLower <- buildStrictPrec(lower)
-        strictOps <- TraverseHelper.traverse(lOps.ops.asInstanceOf[List[LazyParsley[Op]]])(op => suspend(op.optimised[M, R, Op]))
+        strictOps <- traverse(lOps.ops.asInstanceOf[List[LazyParsley[Op]]])(op => suspend(op.optimised[M, R, Op]))
       } yield backend.StrictPrec.Level(strictLower, backend.StrictOps(lOps.f)(strictOps))
     }
   }
